requirements.txt
django~=3.2
python-telegram-bot~=13.13
python-dotenv~=0.20.0
requests~=2.28.1